Reverted changes

+++ 0x0C-more_malloc_free/main.h
@@ -12,7 +12,7 @@ char *string_nconcat(char *s1, char *s2, unsigned int n);
 char *_strncat(char *dest, char *src, int n);
 
 void *_calloc(unsigned int nmemb, unsigned int size);
-void *_memset(void *str, int b, size_t n, size_t size);
+void *_memset(void *str, int b, unsigned int n, unsigned int size);
 
 int *array_range(int min, int max);
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size);
